Update auth Bloc and UI
